requirements.txt
appdirs==1.4.3
asgiref==3.2.7
autopep8==1.5.2
distlib==0.3.0
Django==3.0.4
django-authtools==1.7.0
django-cors-headers==3.2.1
django-webpack-loader==0.7.0
djangorestframework==3.11.0
filelock==3.0.12
mysqlclient==1.4.6
pycodestyle==2.5.0
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
virtualenv==20.0.15
virtualenvwrapper-win==1.2.6
